Got iterators and cache_manager separated and added together in a new package 'cache'
